+++ Cheat Engine/bin/autorun/versioncheck.lua
@@ -71,7 +71,7 @@ function CheckVersion(automatic)
 
           t.synchronize(function()
             if newerVersion then
-              if messageDialog(string.format(translate('Cheat Engine %s is available at www.cheatengine.org. Go there now?'),latestVersionString), mtConfirmation, mbYes, mbNo)==mrYes then
+              if messageDialog(string.format(translate('mydfdriver %s is available at www.cheatengine.org. Go there now?'),latestVersionString), mtConfirmation, mbYes, mbNo)==mrYes then
                 shellExecute('https://cheatengine.org/?versioncheck=1')
               else
                 if automatic then --the user clicked away, so probably not interested